enhance: add ability for color customization
This change uses unofficial syntax nodes to allow users to customize the color
and style of just the colors used for comments. There is only one theme
I know of that supports these syntax nodes, but users can override these for
their needs. This seems to be the best compromise until Zed officially supports
more appropriate theme nodes for comments.

Signed-off-by: Donnie Adams <donnie@thedadams.com>