fix: theme colon and user information the same as the tag (#34)
Previously, in a comment like `TODO(thedadams): something`, then parenthesis,
user, and colon were all style the exact same for all comment types.
Additionally, it was not possible to style the parenthesis and colon without
changing the look of actual code.

This change will default to theming the parenthesis, user, and colon the same as
the rest of the comment. It makes it possible to theme the parenthesis and user
differently for each comment type. The colon is now part of the "text" node and
is themed accordingly.

Signed-off-by: Donnie Adams <donnie@thedadams.com>